+++ lib/strong_migrations/migration.rb
@@ -9,6 +9,9 @@ module StrongMigrations
     def method_missing(method, *args)
       return super if is_a?(ActiveRecord::Schema)
 
+      # Rails 7.0.2+ versioned schema
+      return super if self.class.name.nil?
+
       strong_migrations_checker.perform(method, *args) do
         super
       end