+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.8.0 (unreleased)
+## 0.8.0 (2022-02-09)
 
 - Fixed error with versioned schema with Active Record 7.0.2+
 - Dropped support for Ruby < 2.6 and Active Record < 5.2